+++ Domain/Projects/Project_Development.cs
@@ -45,7 +45,7 @@ namespace Domain.Projects
         public long ProposedPrice { get; set; }
         //مشارکت
         //میزان مشارکت مورد نیاز
-        public int AmountOfParticipationRequired { get; set; }
+        public long AmountOfParticipationRequired { get; set; }
         //خالص ارزش فعلی پروژه
         public long NetPresentValue { get; set; }
