+++ WebApi/ModelsAndDtoes/Project_Development/CreateProject_DevelopmentApiDto.cs
@@ -65,7 +65,7 @@ namespace WebApi.ModelsAndDtoes.Project_Development
         public long ProposedPrice { get; set; }
         //مشارکت
         //میزان مشارکت مورد نیاز
-        public int AmountOfParticipationRequired { get; set; }
+        public long AmountOfParticipationRequired { get; set; }
         //خالص ارزش فعلی پروژه
         public long NetPresentValue { get; set; }
